Changed name dictionary from create on every activation to being stored in dynamodb
Re-implemented ability for 'reviewers' to be displayed
when calling list all issues

Added update messages for listing issues for a user

+++ wxt_cidrbot/dynamo_api_handler.py
@@ -19,20 +19,22 @@ class dynamoapi:
         self.table = self.dynamodb.Table('cidrbot_user_message_preferences')
 
     # Start connection to dynamo. This is done as a function and not in the init because dynamo isn't necessarily called each time the lambda function is run
-    def dynamo_db(self, request, name, status, person_id):
+    def dynamo_db(self, request, name, status, person_id, full_name):
         self.get_dynamo()
         name = self.clean_username(name)
 
         if request == "repos":
             return self.get_repositories()
         if request == "all_users":
-            return self.get_all_users()
+            return self.user_dict()
+        if request == "notif_users":
+            return self.get_notif_users()
         if request == "user_info":
             return self.get_user_info(name)
         if request == "create_user":
-            return self.create_user(name, status, person_id)
+            return self.create_user(name, status, person_id, full_name)
         if request == "update_user":
-            return self.update_user(name, status, person_id)
+            return self.update_user(name, status, person_id, full_name)
         if request == "delete_user":
             return self.delete_user(name)
         if request == "add repo":
@@ -47,6 +49,11 @@ class dynamoapi:
         if "@cisco.com" in name:
             name = name.split("@cisco.com")[0]
         return name
+
+    def user_dict(self):
+        all_users = self.table.scan()
+
+        return all_users
 
     def get_repositories(self):
         self.table = self.dynamodb.Table('cidrbot_repos')
@@ -73,10 +80,11 @@ class dynamoapi:
 
         if db_repo_name == name:
             return f"Repo: {name} already exists"
+
         self.table.put_item(Item={'Repositories': name})
         return f"Successfuly added repo: {name}"
 
-    def get_all_users(self):
+    def get_notif_users(self):
         response = self.table.scan(FilterExpression=Attr('reminders_enabled').eq('on'))
 
         return response
@@ -86,14 +94,36 @@ class dynamoapi:
 
         return response
 
-    def create_user(self, name, status, person_id):
-        self.table.put_item(Item={'User': name, 'person_id': person_id, 'reminders_enabled': status})
+    def create_user(self, name, status, person_id, full_name):
+        first_name = full_name.split(" ")[0].lower()
 
-    def update_user(self, name, status, person_id):
+        response = self.table.scan()
+
+        dup_status = False
+        for user in response['Items']:
+            if first_name == user['first_name']:
+                dup_status = True
+                self.table.update_item(
+                    Key={'User': user['User']},
+                    UpdateExpression='SET dup_status = :dup',
+                    ExpressionAttributeValues={':dup': True}
+                )
+
+        self.table.put_item(
+            Item={
+                'User': name,
+                'person_id': person_id,
+                'reminders_enabled': status,
+                'first_name': first_name,
+                'dup_status': dup_status
+            }
+        )
+
+    def update_user(self, name, status, person_id, full_name):
         response = self.get_user_info(name)
         self.logging.debug(status)
         if len(response['Items']) == 0:
-            self.create_user(name, status, person_id)
+            self.create_user(name, status, person_id, full_name)
         else:
             self.table.update_item(
                 Key={'User': name},
@@ -113,4 +143,26 @@ class dynamoapi:
         return f"Reminders disabled for user: {name}"
 
     def delete_user(self, name):
-        self.table.delete_item(Key={'User': name})
+        # Remove dup status if only 1 person remains with it
+        response = self.table.query(KeyConditionExpression=Key('User').eq(name))
+
+        if len(response['Items']) > 0:
+            self.table.delete_item(Key={'User': name})
+
+        all_users = self.table.scan()
+        dup_counter = 0
+
+        user = ''
+        for user in all_users['Items']:
+            if response['Items'][0]['first_name'] == user['first_name']:
+                if response['Items'][0]['User'] != user['User']:
+                    dup_counter += 1
+
+        # Change dup status to False if 1 user remains
+        if len(user) > 0:
+            if 0 < dup_counter < 2:
+                self.table.update_item(
+                    Key={'User': user['User']},
+                    UpdateExpression='SET dup_status = :dup',
+                    ExpressionAttributeValues={':dup': False}
+                )
